requirements.txt
streamlit
pillow
cryptography
numpy
